requirements.txt
# NL2SQL Chatbot Dependencies
# ============================================================================
# Core LangChain stack for agent orchestration
langchain>=0.3.0
langchain-community>=0.3.0
langchain-databricks>=0.1.0

# LangGraph for ReAct agent pattern
langgraph>=0.2.0,<1.0.0

# Databricks SDK for Unity Catalog and workspace access
databricks-sdk>=0.35.0

# Databricks SQL connector for SQLDatabase.from_databricks()
databricks-sql-connector>=3.0.0

# SQL parsing for security validation
sqlparse>=0.5.0

# Streamlit for chat UI
streamlit>=1.40.0

# Optional: For local development and testing
# python-dotenv>=1.0.0
